requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.3
matplotlib>=3.4.3
streamlit>=1.10.0
numpy>=1.21.2
pillow>=8.3.1